重构 SSH 模块边界
